refactor(runtime): split health and Android config responsibilities
